+++ src/backend/oracle-data-api/src/main/java/ca/bc/gov/open/jag/tco/oracledataapi/repository/JJDisputeRepository.java
@@ -8,7 +8,6 @@ import ca.bc.gov.open.jag.tco.oracledataapi.model.JJDisputeCourtAppearanceAPP;
 import ca.bc.gov.open.jag.tco.oracledataapi.model.JJDisputeCourtAppearanceDATT;
 import ca.bc.gov.open.jag.tco.oracledataapi.model.JJDisputeStatus;
 import ca.bc.gov.open.jag.tco.oracledataapi.model.YesNo;
-import ca.bc.gov.open.jag.tco.oracledataapi.model.TicketImageDataJustinDocument;
 
 public interface JJDisputeRepository {
 
@@ -51,9 +50,6 @@ public interface JJDisputeRepository {
 	/** Fetch all records that match by JJDispute.ticketNumber (should only ever be one). */
 	public List<JJDispute> findByTicketNumber(String ticketNumber);
 	
-	/** Get a justin document record given ticket number and type */
-	public TicketImageDataJustinDocument getTicketImageByRccId(String rccId, String reportType);
-
 	/**
 	 * Saves an entity and flushes changes instantly.
 	 *
