+++ src/backend/oracle-data-api/src/main/java/ca/bc/gov/open/jag/tco/oracledataapi/repository/impl/ords/JJDisputeRepositoryImpl.java
@@ -20,20 +20,16 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Repository;
 
 import ca.bc.gov.open.jag.tco.oracledataapi.mapper.JJDisputeMapper;
-import ca.bc.gov.open.jag.tco.oracledataapi.mapper.TicketImageDataMapper;
 import ca.bc.gov.open.jag.tco.oracledataapi.model.JJDispute;
 import ca.bc.gov.open.jag.tco.oracledataapi.model.JJDisputeCourtAppearanceAPP;
 import ca.bc.gov.open.jag.tco.oracledataapi.model.JJDisputeCourtAppearanceDATT;
 import ca.bc.gov.open.jag.tco.oracledataapi.model.JJDisputeStatus;
-import ca.bc.gov.open.jag.tco.oracledataapi.model.TicketImageDataJustinDocument;
 import ca.bc.gov.open.jag.tco.oracledataapi.model.YesNo;
 import ca.bc.gov.open.jag.tco.oracledataapi.ords.occam.api.handler.ApiException;
 import ca.bc.gov.open.jag.tco.oracledataapi.ords.tco.api.JjDisputeApi;
 import ca.bc.gov.open.jag.tco.oracledataapi.ords.tco.api.model.DisputeResponseResult;
 import ca.bc.gov.open.jag.tco.oracledataapi.ords.tco.api.model.JJDisputeListResponse;
 import ca.bc.gov.open.jag.tco.oracledataapi.ords.tco.api.model.ResponseResult;
-import ca.bc.gov.open.jag.tco.oracledataapi.ords.tco.api.model.TicketImageDataGetParms;
-import ca.bc.gov.open.jag.tco.oracledataapi.ords.tco.api.model.TicketImageDataGetResponseResult;
 import ca.bc.gov.open.jag.tco.oracledataapi.repository.JJDisputeRepository;
 import ca.bc.gov.open.jag.tco.oracledataapi.util.DateUtil;
 
@@ -50,9 +46,6 @@ public class JJDisputeRepositoryImpl implements JJDisputeRepository {
 	@Autowired
 	private JJDisputeMapper jjDisputeMapper;
 	
-	@Autowired
-	private TicketImageDataMapper ticketImageDataMapper;
-
 	public JJDisputeRepositoryImpl(JjDisputeApi jjDisputeApi) {
 		this.jjDisputeApi = jjDisputeApi;
 	}
@@ -111,24 +104,6 @@ public class JJDisputeRepositoryImpl implements JJDisputeRepository {
 		return new ArrayList<JJDispute>();
 	}
 	
-	@Override
-	public TicketImageDataJustinDocument getTicketImageByRccId(String rccId, String reportType) {
-		
-		// set up parameters to pass in
-		TicketImageDataGetParms parms = new TicketImageDataGetParms();
-		ca.bc.gov.open.jag.tco.oracledataapi.ords.tco.api.model.TicketImageDataDocumentKey docKey = new ca.bc.gov.open.jag.tco.oracledataapi.ords.tco.api.model.TicketImageDataDocumentKey(); 
-		docKey.setRccId(rccId);
-		docKey.setReportTypes(reportType);
-		parms.addDocumentKeysItem(docKey); // expects an array of document keys
-
-		TicketImageDataGetResponseResult response = jjDisputeApi.v1TicketImageDataGetPost(parms);
-		if (response == null || response.getDocuments().isEmpty())
-			return new TicketImageDataJustinDocument();
-		
-		TicketImageDataJustinDocument mappedResponse = ticketImageDataMapper.convert(response.getDocuments().get(0));
-		return mappedResponse;
-	}
-
 	@Override
 	public JJDispute saveAndFlush(JJDispute jjDispute) {
 		try {
